// frontend/generated/models/Swipe.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pawdopt api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Details of each Swipe
 * @export
 * @interface Swipe
 */
export interface Swipe {
    /**
     * 
     * @type {string}
     * @memberof Swipe
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Swipe
     */
    adopterId: string;
    /**
     * 
     * @type {string}
     * @memberof Swipe
     */
    dogId: string;
    /**
     * 
     * @type {string}
     * @memberof Swipe
     */
    direction: SwipeDirectionEnum;
    /**
     * 
     * @type {string}
     * @memberof Swipe
     */
    shelterId: string;
    /**
     * 
     * @type {Date}
     * @memberof Swipe
     */
    swipedAt: Date;
}


/**
 * @export
 */
export const SwipeDirectionEnum = {
    Left: 'left',
    Right: 'right'
} as const;
export type SwipeDirectionEnum = typeof SwipeDirectionEnum[keyof typeof SwipeDirectionEnum];


/**
 * Check if a given object implements the Swipe interface.
 */
export function instanceOfSwipe(value: object): value is Swipe {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('adopterId' in value) || value['adopterId'] === undefined) return false;
    if (!('dogId' in value) || value['dogId'] === undefined) return false;
    if (!('direction' in value) || value['direction'] === undefined) return false;
    if (!('shelterId' in value) || value['shelterId'] === undefined) return false;
    if (!('swipedAt' in value) || value['swipedAt'] === undefined) return false;
    return true;
}

export function SwipeFromJSON(json: any): Swipe {
    return SwipeFromJSONTyped(json, false);
}

export function SwipeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Swipe {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'adopterId': json['adopterId'],
        'dogId': json['dogId'],
        'direction': json['direction'],
        'shelterId': json['shelterId'],
        'swipedAt': (new Date(json['swipedAt'])),
    };
}

export function SwipeToJSON(json: any): Swipe {
    return SwipeToJSONTyped(json, false);
}

export function SwipeToJSONTyped(value?: Swipe | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'adopterId': value['adopterId'],
        'dogId': value['dogId'],
        'direction': value['direction'],
        'shelterId': value['shelterId'],
        'swipedAt': ((value['swipedAt']).toISOString()),
    };
}