// frontend/generated/models/SwipeCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pawdopt api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Details of each Swipe
 * @export
 * @interface SwipeCreate
 */
export interface SwipeCreate {
    /**
     * 
     * @type {string}
     * @memberof SwipeCreate
     */
    dogId: string;
    dogCreatedAt: string;
    /**
     * 
     * @type {string}
     * @memberof SwipeCreate
     */
    shelterId: string;
    /**
     * 
     * @type {string}
     * @memberof SwipeCreate
     */
    direction: SwipeCreateDirectionEnum;
}


/**
 * @export
 */
export const SwipeCreateDirectionEnum = {
    Left: 'left',
    Right: 'right'
} as const;
export type SwipeCreateDirectionEnum = typeof SwipeCreateDirectionEnum[keyof typeof SwipeCreateDirectionEnum];


/**
 * Check if a given object implements the SwipeCreate interface.
 */
export function instanceOfSwipeCreate(value: object): value is SwipeCreate {
    if (!('dogId' in value) || value['dogId'] === undefined) return false;
    if (!('direction' in value) || value['direction'] === undefined) return false;
    return true;
}

export function SwipeCreateFromJSON(json: any): SwipeCreate {
    return SwipeCreateFromJSONTyped(json, false);
}

export function SwipeCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): SwipeCreate {
    if (json == null) {
        return json;
    }
    return {
        'dogId': json['dog_id'],
        'dogCreatedAt': json['dog_created_at'],
        'direction': json['direction'],
        'shelterId': json['shelter_id']
    };
}

export function SwipeCreateToJSON(json: any): SwipeCreate {
    return SwipeCreateToJSONTyped(json, false);
}

export function SwipeCreateToJSONTyped(value?: SwipeCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }
    console.log("value:", value)
    return {
        'dogId': value['dogId'],
        'dogCreatedAt': value['dogCreatedAt'],
        'direction': value['direction'],
        'shelterId': value['shelterId']
    };
}